requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
certifi==2022.6.15
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via flask
colorama==0.4.5
    # via click
flask==2.1.2
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
importlib-metadata==4.11.4
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.1
    # via jinja2
python-dotenv==0.20.0
    # via -r requirements.in
requests==2.27.1
    # via -r requirements.in
urllib3==1.26.9
    # via requests
werkzeug==2.1.2
    # via flask
zipp==3.8.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
